requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.1.2
pytesseract==0.3.4
Werkzeug==1.0.1
